Redevelopped a lot of things cleaner
Cleaning the code, redevelopping things, and make stuff really modular.
Now things are made in proper modules with settable arguments, meaning
that it's easy to have multiple variants of the wheels in the same
design.
Added new functionnalities, such as countersunk holes for screw, and the
ability to have trough hole or partial holes for screws, on both faces
of the wheel.
Holes and arms can also be offset from any angle (each have a separate
setting)