added a functionnal tank track module
Here, i added a tank track module (also parametric) that matches the
wheels. It has been tweeked for better fit on the wheels, and has a
simple outer gripping pattern. It's meant to be printed in soft, rubbery
plastics as ninjaflex or semiflex.
A lot of STL files have also been added; in this release, only the 12
and 16 bumps models are tweaked; the 29 have been tested, but needs
tweaking.